Fold the TTS chat message into the donation renderer
TTS is no longer a distinct event: a read-aloud donation is a DONATION
carrying an audio URL. Render it with ChatDonationMessage, adding the
click-to-play button when audio is present, and drop ChatTTSMessage,
the TTS message type, builder, event-bar cases, and _tts.scss (its
play-button styles move onto .msg-donation). The click handler and
audio playback stay — they key off the .tts-play class, not the type.
