+++ assets/chat/css/messages/event/_event.scss
@@ -1,7 +1,6 @@
 @use '../../abstracts/' as a;
 
 .msg-donation,
-.msg-tts,
 .msg-subscription,
 .msg-giftsub,
 .msg-massgift,
