+++ assets/chat/css/messages/event/_index.scss
@@ -4,7 +4,6 @@
 
 @use 'broadcast';
 @use 'donation';
-@use 'tts';
 @use 'subscription';
 @use 'xpost';
 @use 'vestaboard';
