--- assets/chat/css/messages/event/_tts.scss
@@ -1,33 +1,0 @@
-@use '../../abstracts/' as a;
-
-// A TTS event renders exactly like a donation (see _donation and _event); the
-// only addition is the "Play TTS" button stacked below the message.
-.msg-tts {
-  .event-bottom {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 0.25rem;
-  }
-
-  .event-bottom-action {
-    display: inline-flex;
-    align-items: center;
-    margin-bottom: 0.25rem;
-    padding: 0.25em a.$gutter-md;
-    border: none;
-    border-radius: 6px;
-    background-color: #2a2a2a;
-    color: a.$color-chat-text1;
-    text-decoration: none;
-    font-weight: 600;
-    line-height: 1;
-    cursor: pointer;
-
-    &:hover,
-    &:focus-visible {
-      background-color: #353535;
-      color: a.$color-chat-text1;
-    }
-  }
-}
