+++ assets/chat/js/chat.js
@@ -160,7 +160,6 @@ class Chat {
     this.source.on('GIFTSUB', (data) => this.onGIFTSUB(data));
     this.source.on('MASSGIFT', (data) => this.onMASSGIFT(data));
     this.source.on('DONATION', (data) => this.onDONATION(data));
-    this.source.on('TTS', (data) => this.onTTS(data));
     this.source.on('XPOST', (data) => this.onXPOST(data));
     this.source.on('VESTABOARD_LEAD', (data) => this.onVESTABOARD_LEAD(data));
     this.source.on('VESTABOARD_HOURLY', (data) =>
@@ -906,7 +905,6 @@ class Chat {
         MessageTypes.GIFTSUB,
         MessageTypes.MASSGIFT,
         MessageTypes.DONATION,
-        MessageTypes.TTS,
         MessageTypes.BROADCAST,
         MessageTypes.DEATH,
       ].includes(message.type)
@@ -1632,18 +1630,6 @@ class Chat {
     audio.addEventListener('ended', stopped);
     audio.addEventListener('error', stopped);
     audio.play().catch(stopped);
-  }
-
-  onTTS(data) {
-    MessageBuilder.tts(data).into(this);
-
-    if (!this.backlogloading) {
-      const eventBarEvent = new EventBarEvent(this, MessageTypes.TTS, data);
-      this.eventBar.add(eventBarEvent);
-      if (this.eventBar.length === 1) {
-        this.mainwindow.update();
-      }
-    }
   }
 
   onXPOST(data) {
