+++ assets/chat/js/messages/MessageTypes.js
@@ -13,7 +13,6 @@ export default {
   GIFTSUB: 'GIFTSUB',
   MASSGIFT: 'MASSGIFT',
   DONATION: 'DONATION',
-  TTS: 'TTS',
   XPOST: 'XPOST',
   DEATH: 'DEATH',
   VESTABOARD_LEAD: 'VESTABOARD_LEAD',
