+++ assets/chat/js/messages/index.js
@@ -2,7 +2,6 @@ export { default as MessageBuilder } from './MessageBuilder';
 export { default as MessageTypes } from './MessageTypes';
 export { default as ChatMessage } from './ChatMessage';
 export { default as ChatDonationMessage } from './ChatDonationMessage';
-export { default as ChatTTSMessage } from './ChatTTSMessage';
 export { default as ChatXPostMessage } from './ChatXPostMessage';
 export { default as ChatEmoteMessage } from './ChatEmoteMessage';
 export { default as ChatUIMessage } from './ChatUIMessage';
